+++ src/settings/ServicePoints/utils.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
+import { get } from 'lodash';
 
 export const validateServicePointForm = (values) => {
   const errors = {};
@@ -54,10 +55,14 @@ export const getUniquenessValidation = (field, mutator) => {
   };
 };
 
-export const isEcsRequestRoutingVisible = (stripes) => (
-  stripes.hasInterface('consortia') && stripes.hasInterface('ecs-tlr')
+export const isEcsRequestRoutingVisible = (titleLevelRequestsFeatureEnabled) => (
+  !!titleLevelRequestsFeatureEnabled
 );
 
-export const isEcsRequestRoutingAssociatedFieldsVisible = (stripes, ecsRequestRouting) => (
-  (isEcsRequestRoutingVisible(stripes) && !ecsRequestRouting) || !isEcsRequestRoutingVisible(stripes)
+export const isEcsRequestRoutingAssociatedFieldsVisible = (titleLevelRequestsFeatureEnabled, ecsRequestRouting) => (
+  (isEcsRequestRoutingVisible(titleLevelRequestsFeatureEnabled) && !ecsRequestRouting) || !isEcsRequestRoutingVisible(titleLevelRequestsFeatureEnabled)
+);
+
+export const getEcsTlrFeature = (data = []) => (
+  get(data, '[0].value.enabled', false)
 );